+++ CHANGELOG.rst
@@ -69,6 +69,7 @@ Changed
 - feat: added `to_json` method for ToDatetimeTransformer `#553 <https://github.com/azukds/tubular/issues/553>`_
 - feat: added 'to_json' method for OneDKmeansTransformer `#551 <https://github.com/azukds/tubular/issues/551>_`
 - bugfix: Fixed issue with OneHotEncodingTransformer when used with wanted_values and unseen levels at transform time
+- feat: added two new transformers called WhenThenOtherwiseTransformer and CompareTwoColumnsTransformer
 
 2.4.0 (01/12/2025)
 ------------------